refactor(type): Product, ProductSummary 타입을 공통 정의로 통합 활용
- useThemeProducts 및 ThemeProductList에서 중복된 Product 타입을 제거했습니다.
- theme과 관련된 타입도 차후 재사용성을 고려해 theme.ts로 추출해주었습니다.
