+++ src/hooks/useThemeProducts.ts
@@ -1,22 +1,7 @@
 import { useEffect, useState, useCallback, useRef } from 'react';
 import axios from 'axios';
 import { getThemeProductsUrl } from '@/constants/api';
-
-interface Product {
-  id: number;
-  name: string;
-  price: {
-    basicPrice: number;
-    sellingPrice: number;
-    discountRate: number;
-  };
-  imageURL: string;
-  brandInfo: {
-    id: number;
-    name: string;
-    imageURL: string;
-  };
-}
+import type { Product } from '@/types/product';
 
 const LIMIT = 10;
