+++ src/types/theme.ts
@@ -1,0 +1,7 @@
+export interface ThemeInfo {
+  themeId: number;
+  name: string;
+  title: string;
+  description: string;
+  backgroundColor: string;
+}
